+++ crawler_system/include/robot_motion/flipper_util.h
@@ -1,0 +1,15 @@
+#ifndef UTIL_H_
+#define UTIL_H_
+
+#include<cmath>
+
+template<typename T>
+inline T degToRad(const T deg){
+  return deg * (M_PI / 180.0);
+}
+template<typename T>
+inline T radToDeg(const T rad){
+  return rad * (180.0 / M_PI);
+}
+
+#endif
