+++ crawler_system/src/flipper_semi_autonomous.cpp
@@ -16,6 +16,8 @@
 #include "robot_motion/semi_autonomous.h"
 #include "robot_motion/Rotation.h"
 #include "robot_motion/DXL.h"
+#include "robot_motion/Constant.h"
+#include "robot_motion/flipper_util.h"
 //#include "dynamixel/dynamixel.h"
 //#include "dynamixel.cpp"
 #include <dynamixel_workbench_msgs/DynamixelCommand.h>
@@ -23,53 +25,7 @@
 //HACK:全ての変数、関数をnamespaceの中に入れる
 //HACK:グローバル宣言の変数多すぎもっと参照渡し使おう
 //HACK:パラメータ関係を別のファイルにまとめる
-namespace DXL{
-  extern template class DXLControl
-}
-namespace FlipperConstant{
-  //HACK:一つずつサーボIDを管理するのは頭が悪い
-  constexpr int front_right = 0;
-  constexpr int front_left = 1;
-  constexpr int rear_right = 2;
-  constexpr int rear_left = 3;
-
-  constexpr double flipper_m = 100;
-  constexpr double flipper_lg = 100;
-  constexpr double gravity = 9.81;
-};
-namespace DXLConstant{
-  constexpr double ORIGIN_DEG = 90.0;
-  constexpr double AUTO_MAX_DEG = 75.0;
-  constexpr double AUTO_MIN_DEG = -90.0;
-  constexpr int MAX_POSITION_VALUE = 1048575;
-  constexpr int MIN_POSITION_VALUE = -1048575;
-  constexpr int DYNAMIXEL_RESOLUTION = 4096;
-  constexpr double DYNAMIXEL_RESOLUTION_ANGLE = 0.088;
-  constexpr bool TORQUE_ENABLE = 1;
-  constexpr bool TORQUE_DISABLE = 0;
-  //TODO:それぞれのサーボゲインを設ける
-  constexpr double Kp = 1.0;
-  constexpr double Kd = 1.0;
-};
-std::array<double, 4>     ref_DXL_raw_pos{};
-std::array<double, 4> current_DXL_raw_pos{};
-
-std::array<double, 4>     ref_DXL_rad{};
-std::array<double, 4> current_DXL_rad{};
-
-std::array<double, 4>     ref_DXL_torque{};
-std::array<double, 4> current_DXL_torque{};
-
-constexpr std::array<int, 4> dynamixel_num{0, 1, 3, 2};
-
-template<typename T>
-inline T degToRad(const T deg){
-  return deg * (M_PI / 180.0);
-}
-template<typename T>
-inline T radToDeg(const T rad){
-  return rad * (180.0 / M_PI);
-}
+const DXL::MODE DXL_MODE = DXL::MODE::TORQUE_CONTROL;
 
 enum class keyFlag{
   NOMAL = 0,
@@ -78,7 +34,7 @@ enum class keyFlag{
 };
 
 //-----------------------------------------------------------------
-namespace safetyCheck{
+/*namespace safetyCheck{
   //過負荷の確認
   void torque_limit(){
   }
@@ -119,8 +75,7 @@ namespace gyroControl{
     }
   }
 };
-
-constexpr DXL::MODE DXL_MODE = DXL::MODE::TORQUE_CONTROL;
+*/
 
 template<typename T>
 struct Gyro{
@@ -133,6 +88,7 @@ Gyro<double> gyro_robot;
 
 //ロボットの現在角度を取得
 //rosのtfに合わせてあとで型を変える
+/*
 namespace RobotState{
   void gyroCallback(const std_msgs::Float64 &msg) {
     gyro_robot.x = msg.data.x;
@@ -147,10 +103,10 @@ namespace RobotState{
       //配列番号を変える
       std::swap(dynamixel_num[0], dynamixel_num[2]);
       std::swap(dynamixel_num[1], dynamixel_num[3]);
-    }*/
+    }
   }
 };
-
+*/
 bool buttons_reverse = false;
 bool flag_reset = false;
 bool prev_reset = false;
@@ -173,18 +129,58 @@ void joyCallback(const sensor_msgs::Joy &controller) {
   prev_reset = controller.buttons[1];
 }
 
+ros::ServiceClient dynamixel_service;
+dynamixel_workbench_msgs::DynamixelCommand srv;
+
+bool serviceCallPos(){
+  for(int i = 0; i < dynamixel_num.size(); ++i){
+    srv.request.command = "_";
+    srv.request.id = i + 1;
+    srv.request.addr_name = "Goal_Position";
+    srv.request.value = ref_DXL_raw_pos[i];
+    dynamixel_service.call(srv);
+  }
+  //TODO:いるかあとで考える
+  ros::spinOnce();
+  return true;
+}
+bool serviceCallTheta(){
+  for (int i = 0; i < dynamixel_num.size(); ++i) {
+    //TODO:今のままだと計算がバグる
+    if(ref_DXL_rad[i] > 360)ref_DXL_rad[i] -= 360;
+    if(ref_DXL_rad[i] > 0)ref_DXL_rad[i] += 360;
+    ROS_INFO("ref_DXL_rad[%d] %lf current_DXL_rad[%d] %lf", i, ref_DXL_rad[i], i, current_DXL_rad[i]);
+    srv.request.command = "_";
+    srv.request.id = i + 1;
+    srv.request.addr_name = "Goal_Position";
+    srv.request.value = DXL::dynamixelSet(ref_DXL_rad[i], current_DXL_rad[i]);
+    dynamixel_service.call(srv);
+  }
+  ros::spinOnce();
+}
+//現在角度とトルクを取得
+//TODO: dynamixelIDとjointstateの順番が違う
+void jointStateCallback(const sensor_msgs::JointState &jointstate) {
+  for(int i = 0; i < dynamixel_num.size(); ++i){
+    current_DXL_raw_pos[i] = jointstate.position[dynamixel_num[i]];
+    //FIXME:これでは角度を得られない(DXLの位置値を直接変換しているから一度ラジアンにしてdegToRadを使用する必要がある)
+    current_DXL_rad[i] = degToRad<double>((jointstate.position[dynamixel_num[i]] + M_PI));
+    current_DXL_torque[i] = jointstate.effort[dynamixel_num[i]];
+  }
+}
 int main(int argc, char **argv) {
   ros::init(argc, argv, "semi_auto");
   ros::NodeHandle n;
 
   dynamixel_service = n.serviceClient<dynamixel_workbench_msgs::DynamixelCommand>("/dynamixel_workbench/dynamixel_command");
-  ros::Subscriber feedback_sub = n.subscribe("/dynamixel_workbench/joint_states", 10, DXL::jointStateCallback);
-  ros::Subscriber gyro_sub = n.subscribe("gyro", 10, RobotState::gyroCallback);
+  ros::Subscriber feedback_sub = n.subscribe("/dynamixel_workbench/joint_states", 10, jointStateCallback);
+  //ros::Subscriber gyro_sub = n.subscribe("gyro", 10, RobotState::gyroCallback);
   ros::Subscriber controller_sub = n.subscribe("joy", 10, joyCallback);
   ros::Rate loop_rate(400);
   //REVIEW:多分これだとtemplate<DXLConstant::DXL_MODE>のオブジェクトが4つ生成されない??
   
-  std::array<DXL::DXLControl<double, DXL_MODE>, 4> servo{  
+  //std::array<DXL::DXLControl<double, DXL_MODE>, 4> servo{  
+  DXL::DXLControl<double, DXL_MODE> servo[4]{  
     FlipperConstant::front_right, 
     FlipperConstant::front_left, 
     FlipperConstant::rear_right,
